+++ Expensive.Domain/Entities/Entries.cs
@@ -2,12 +2,12 @@
 
 public class Entries : BaseEntity
 {
-    public int EntryId { get; set; }
+    public long EntryId { get; set; }
     public decimal Amount { get; set; }
-    public int UserId { get; set; }
+    public long UserId { get; set; }
     public Users? User { get; set; }
-    public int CategorieId { get; set; }
+    public long CategorieId { get; set; }
     public Categories? Categorie { get; set; }
-    public int PaymentMethodId { get; set; }
+    public long PaymentMethodId { get; set; }
     public PaymentMethods? PaymentMethod { get; set; }
 }
