+++ Expensive.Domain/Entities/UserCategories.cs
@@ -3,8 +3,8 @@
 public class UserCategories : BaseEntity
 {
     public long UserCategorieId { get; set; }
-    public int UserId { get; set; }
+    public long UserId { get; set; }
     public Users? User { get; set; }
-    public int CategorieId { get; set; }
+    public long CategorieId { get; set; }
     public Categories? Categorie { get; set; }
 }
