+++ Expensive.Domain/Entities/UserPaymentMethods.cs
@@ -2,9 +2,9 @@
 
 public class  UserPaymentMethods : BaseEntity
 {
-    public int UserPaymentMethodId { get; set; }
-    public int UserId { get; set; }
+    public long UserPaymentMethodId { get; set; }
+    public long UserId { get; set; }
     public Users? User { get; set; }
-    public int PaymentMethodId { get; set; }
+    public long PaymentMethodId { get; set; }
     public PaymentMethods? PaymentMethod { get; set; }
 }
